backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
channels==4.0.0
channels-redis==4.1.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
uvicorn==0.24.0
daphne==4.0.0
whitenoise==6.6.0
python-dotenv==1.0.0
pandas==2.1.3
pytest==7.4.3
pytest-django==4.5.2
scikit-learn==1.3.2
joblib==1.3.2
xgboost==2.0.2
djangorestframework-simplejwt==5.3.1
reportlab==4.0.7
matplotlib==3.8.2
openpyxl==3.1.2
